+++ cart/views.py
@@ -5,7 +5,7 @@ from rest_framework import status
 
 from .models import Carts
 from .serializers import CartSerializer
-from product.models import Products
+# from product.models import Products
 
 from drf_spectacular.utils import extend_schema
 class CartView(APIView):
@@ -40,10 +40,14 @@ class CartView(APIView):
             return Response(CartSerializer(cart_item).data)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self, request, pk):
+    def delete(self, request, *args, **kwargs):
+        pk = kwargs.get('pk')
+        if not pk:
+            return Response({"error": "ID required for deletion"}, status=status.HTTP_400_BAD_REQUEST)
+
         try:
             cart_item = Carts.objects.get(pk=pk, user=request.user)
             cart_item.delete()
-            return Response({"message": "Item removed from cart"})
+            return Response({"message": "Item removed from cart"}, status=status.HTTP_200_OK)
         except Carts.DoesNotExist:
             return Response({"error": "Item not found"}, status=status.HTTP_404_NOT_FOUND)
